// src/app/models/app-products-update-dto.ts
export class AppProductsUpdateDto {

     id: number;
     Code: string;
     externalCode: string;
     Description1: string;
     Description2: string;
     AppUnitsId: number;
     ProductionUnitId: number;
     UnitPrice: number;
     PrymaryMtrMonedaId: number;
     SecundaryMtrMonedaId: number;
     UrlImage: string;
     AppSubCategoryId: number;
     UsuarioConectado: string;
     quantityPerPackage: number;
     validadCantidadFija: boolean;
     cantidadFija: number;
     cantidadMinima: number;
     tipoCalculo: number;
     requiereEstimacion: boolean;

}